+++ papermerge/core/features/document_types/tests/test_router_document_types.py
@@ -1,7 +1,8 @@
 from sqlalchemy import func
 from sqlalchemy.orm import Session
 
-from papermerge.core import schemas
+from papermerge.core.features.custom_fields import schema as cf_schema
+from papermerge.core.features.document_types import schema as dt_schema
 from papermerge.core.features.document_types.db import DocumentType
 from papermerge.test.types import AuthTestClient
 
@@ -25,7 +26,7 @@ def test_create_document_type_with_path_template(
     count_after = db_session.query(func.count(DocumentType.id)).scalar()
     assert count_after == 1
 
-    document_type = schemas.DocumentType.model_validate(response.json())
+    document_type = dt_schema.DocumentType.model_validate(response.json())
     assert document_type.name == "Invoice"
     assert document_type.path_template == "/home/My ZDF/"
 
@@ -43,15 +44,15 @@ def test_update_document_type_with_path_template(
         },
     )
 
-    document_type = schemas.DocumentType.model_validate(response.json())
+    document_type = dt_schema.DocumentType.model_validate(response.json())
     assert document_type.path_template == "/home/My ZDF/updated/"
 
 
 def test_create_document_type(
     make_custom_field, auth_api_client: AuthTestClient, db_session: Session
 ):
-    cf1: schemas.CustomField = make_custom_field(name="shop", type="text")
-    cf2: schemas.CustomField = make_custom_field(name="total", type="monetary")
+    cf1: cf_schema.CustomField = make_custom_field(name="shop", type="text")
+    cf2: cf_schema.CustomField = make_custom_field(name="total", type="monetary")
 
     count_before = db_session.query(func.count(DocumentType.id)).scalar()
     assert count_before == 0
@@ -66,7 +67,7 @@ def test_create_document_type(
     count_after = db_session.query(func.count(DocumentType.id)).scalar()
     assert count_after == 1
 
-    document_type = schemas.DocumentType.model_validate(response.json())
+    document_type = dt_schema.DocumentType.model_validate(response.json())
     assert document_type.name == "Invoice"
     assert len(document_type.custom_fields) == 2
     assert set([cf.name for cf in document_type.custom_fields]) == {"shop", "total"}
@@ -97,7 +98,7 @@ def test_update_document_type(
         },
     )
     assert response.status_code == 200
-    updated_dtype = schemas.DocumentType(**response.json())
+    updated_dtype = dt_schema.DocumentType(**response.json())
     assert updated_dtype.name == "Invoice-updated"
     assert set([cf.name for cf in updated_dtype.custom_fields]) == {"cf1", "cf2"}
